+++ Snake/Snake/NormalFood.cpp
@@ -12,7 +12,7 @@ NormalFood::~NormalFood()
 {
 }
 
-void NormalFood::eatingFood(std::vector<Snake> &snake, int &points)
+void NormalFood::eatingFood(std::vector<Snake> &snake, unsigned long int &points)
 {
 	snake.push_back(Snake(snake.back().getX(), snake.back().getY()));
 	points += 10;
