+++ Snake/Snake/NormalFood.h
@@ -9,6 +9,6 @@ class NormalFood :
 public:
 	NormalFood();
 	~NormalFood();
-	void eatingFood(std::vector<Snake> &snake, int &points);
+	void eatingFood(std::vector<Snake> &snake, unsigned long int &points);
 };
